Add files via upload
We used the NNAR and linear regression withARIMA methods learned in the course to compare the effects on this data. As an additional try, add an MLP to add the x variable, and then combine it with NNAR to try to add the influence of the x variable to the NNAR model.